Fix bug in trimGlobs.
The algorithm can yield an incorrect result when one of the globs
is trimmed until it's empty. Example:
NonEmpty("foo.*bar", "foobar") will be reduced to
NonEmpty(".*", "") which will yield `false`, which is incorrect
for the original input.

The fix leaves exactly one token untrimmed on both ends, so the
same input will now be reduced to NonEmpty("o.*b", "ob") which
yields the correct output.