Fix for `[ab]+c ~ [bc]+` cases
First of, thanks for this package!  I was looking at some weird behaviour in OPA
rules and the rabbit hole ended up leading here.

The case `[ab]+c ~ [bc]+` was the smallest I was able to reproduce it down to.
The string `bc` is matched by both regexes, so there should be an intersection,
however `NonEmpty` returns `false`.

Digging into the code a bit, if we're comparing a plussed token on the left
hand side, and the first token on the right hand side matches, it seems
we always end up dropping that first token on the right hand side before
continuing.

This is not always acceptable though, consider the example `[ab]+c ~ [bc]+`. we
end up comparing `c` with the empty string in the second step.

In order to account for this, I introduced a second branch for the cases where
the first token on the right hand side maybe completely consume the the first
token on the left hand side (without eliminating itself).  This is only possible
if the token on the right hand side also has `+` or `*`, so AFAIK this doesn't
change the theoretical complexity.